backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
langgraph>=0.2.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.7
httpx>=0.27.0
python-multipart==0.0.9
supabase>=2.4.0
python-dotenv==1.0.1
Pillow>=10.3.0
